Backened/requirements.txt
annotated-types==0.7.0
anyio==4.11.0
arabic-reshaper==3.0.0
asn1crypto==1.5.1
blinker==1.9.0
certifi==2025.8.3
cffi==2.0.0
charset-normalizer==3.4.3
click==8.3.0
colorama==0.4.6
cryptography==46.0.2
cssselect2==0.8.0
dnspython==2.8.0
easyocr==1.7.2
ecdsa==0.19.1
email-validator==2.3.0
fastapi==0.118.0
fastapi-cli==0.0.13
fastapi-cloud-cli==0.2.1
ffmpeg-python==0.2.0
filelock==3.19.1
Flask==3.1.2
flask-cors==6.0.1
freetype-py==2.5.1
fsspec==2025.9.0
future==1.0.0
h11==0.16.0
html5lib==1.1
httpcore==1.0.9
httptools==0.6.4
httpx==0.28.1
idna==3.10
imageio==2.37.0
itsdangerous==2.2.0
Jinja2==3.1.6
jiwer==4.0.0
joblib==1.5.2
lazy_loader==0.4
llvmlite==0.45.0
lxml==6.0.2
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mdurl==0.1.2
more-itertools==10.8.0
mpmath==1.3.0
networkx==3.5
ninja==1.13.0
numba==0.62.1
numpy==2.2.6
openai-whisper==20250625
opencv-python-headless==4.12.0.88
orjson==3.11.3
oscrypto==1.3.0
packaging==25.0
pandas==2.3.3
pillow==11.3.0
pip==24.0
pyasn1==0.6.1
pycairo==1.28.0
pyclipper==1.3.0.post6
pycparser==2.23
pydantic==2.11.9
pydantic_core==2.33.2
pydantic-extra-types==2.10.5
pydantic-settings==2.11.0
Pygments==2.19.2
pyHanko==0.31.0
pyhanko-certvalidator==0.29.0
pypdf==6.1.1
pyphen==0.17.2
python-bidi==0.6.6
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-jose==3.5.0
python-multipart==0.0.20
pytz==2025.2
PyYAML==6.0.3
RapidFuzz==3.14.1
regex==2025.9.18
reportlab==4.4.4
requests==2.32.5
rich==14.1.0
rich-toolkit==0.15.1
rignore==0.6.4
rlPyCairo==0.4.0
rsa==4.9.1
scikit-image==0.25.2
scikit-learn==1.7.0
scipy==1.16.2
sentry-sdk==2.39.0
setuptools==65.5.0
shapely==2.1.2
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
starlette==0.48.0
svglib==1.6.0
sympy==1.14.0
threadpoolctl==3.6.0
tifffile==2025.9.30
tiktoken==0.11.0
tinycss2==1.4.0
torch==2.8.0
torchvision==0.23.0
tqdm==4.67.1
typer==0.19.2
typing_extensions==4.15.0
typing-inspection==0.4.2
tzdata==2025.2
tzlocal==5.3.1
ujson==5.11.0
uritools==5.0.0
urllib3==2.5.0
uvicorn==0.37.0
watchfiles==1.1.0
webencodings==0.5.1
websockets==15.0.1
Werkzeug==3.1.3
whisper==1.1.10
xhtml2pdf==0.2.17
# python-socketio==5.11.2
# python-engineio==4.9.0
# socketio==0.2.1